Rename DisconnectSite component to DisconnectSiteLink
Current naming will be in conflict with the name
of the Disconnect Site in the new disconnect flow.
